+++ packages/replay/test/types.ts
@@ -1,4 +1,1 @@
-import { Transport } from '@sentry/types';
-
-export type MockTransportSend = jest.MockedFunction<Transport['send']>;
 export type DomHandler = (args: any) => any;